design: Header 파일 위치 변경, 디자인
